fix: switch AMP from fp16 to bfloat16 - no GradScaler needed
FP16 autocast caused persistent gradient overflow (nan grad_norm),
meaning optimizer never stepped. BF16 has same dynamic range as fp32
(8 exponent bits) so never overflows. H20 natively supports BF16.

Removed GradScaler (unnecessary for BF16). Keep CovInv as float32
since local_aggregate CUDA kernel requires it.

+++ model/head/gaussian_head.py
@@ -89,7 +89,7 @@ class GaussianHead(BaseTaskHead):
         R = get_rotation_matrix(rotations) # b, g, 3, 3
         M = torch.matmul(S, R)
         Cov = torch.matmul(M.transpose(-1, -2), M)
-        CovInv = Cov.float().cpu().inverse().cuda().to(Cov.dtype) # b, g, 3, 3; float() for AMP compatibility
+        CovInv = Cov.float().cpu().inverse().cuda() # b, g, 3, 3; keep float32 (local_aggregate needs fp32)
         return means, origi_opa, opacities, scales, CovInv
 
     def get_filtered_lidar(self,lidar):
